experiments/requirements.txt
bertopic==0.15.0
sentence-transformers>=2.6.0
datasets==2.16.1
evaluate==0.4.1
torch>=2.0.0
transformers>=4.36.0
pandas>=2.0.0
gensim>=4.3.0
tqdm>=4.65.0
numpy>=1.23.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
hf_xet>=0.1.0
huggingface-hub>=0.21.0
bitsandbytes>=0.41.0 